+++ inst/include/model_helpers.hpp
@@ -55,9 +55,9 @@ namespace emphasis {
   {
     double brts;
     double n;         /* n[i] = number of species in [time_i-1, time_i) */
-  double t_ext;     /* emp_t_ext_tip for present-day species;  emp_t_ext_extinct for extinction nodes */
-  double pd;
-  int clade; // required for sim_tree
+    double t_ext;     /* emp_t_ext_tip for present-day species;  emp_t_ext_extinct for extinction nodes */
+    double pd;
+    int clade;        // required for sim_tree
   };
