+++ src/rcpp_mce.cpp
@@ -7,27 +7,9 @@
 #include "emphasis.hpp"
 #include "model.hpp"
 #include "rinit.h"
+#include "unpack.h"
 using namespace Rcpp;
 
-
-namespace {
-
-  DataFrame unpack(const emphasis::tree_t& tree)
-  {
-    NumericVector brts, n, t_ext, pd;
-    for (const emphasis::node_t& node : tree) {
-      brts.push_back(node.brts);
-      n.push_back(node.n);
-      t_ext.push_back(node.t_ext);
-      pd.push_back(node.pd);
-    }
-    return DataFrame::create(Named("brts") = brts, 
-                             Named("n") = n, 
-                             Named("t_ext") = t_ext,
-                             Named("pd") = pd);
-  }
-
-}
 
 //' function to perform one step of the E-M algorithm
 //' @param brts vector of branching times
